+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+1.1.0
+-----
+
+* Highlight GPIO label if `--find` matches a hidden label
+* Add regex support to `--find`
+* Rewrite render flow to make it less hacky
+* Light mode is back! `--light`
+* Move error text above usage message
+
 1.0.1
 -----